make TcpProxyConfig an interface
