+++ test/com/github/terma/javaniotcpproxy/TcpProxyTest.java
@@ -4,7 +4,7 @@ import org.junit.Test;
 
 public class TcpProxyTest {
 
-    private TcpProxyConfig config = new TcpProxyConfig(0, "", 0);
+    private StaticTcpProxyConfig config = new StaticTcpProxyConfig(0, "", 0);
 
     @Test(expected = IllegalArgumentException.class)
     public void shouldFailWhenCreateWithNegativeWorkersCount() {
